+++ Engine/Recources/ShaderManager.cs
@@ -18,7 +18,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
-using HelperSuite.ContentLoader;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;